+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- Improved METAR output
+
 ## [2.4.0] 2026-09-05
 
 - Changed organisation of default application constants